refactor: 상세 페이지 쿼리를 따로 관리
