+++ src/app/(detail)/containers/ProductDetailContainer/ProductDetailContainer.tsx
@@ -4,7 +4,7 @@ import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 
 import { getQueryClient } from '@/app/(configs)/query/config';
 import GBItemDetail from '@/app/(detail)/components/GBItemDetail';
-import { getProductsDetailQueryObject } from '@/app/(queries)/productsQueries';
+import { getProductDetailQueryObject } from '@/app/(queries)/productQueries';
 
 type ProductDetailContainerProps = {
   productId: string;
@@ -14,7 +14,7 @@ const ProductDetailContainer = async ({ productId }: ProductDetailContainerProps
   const queryClient = getQueryClient();
 
   try {
-    await queryClient.prefetchQuery(getProductsDetailQueryObject(productId));
+    await queryClient.prefetchQuery(getProductDetailQueryObject(productId));
   } catch (error) {
     console.error(error);
   }
