+++ src/app/(queries)/productsQueries.ts
@@ -1,13 +1,7 @@
 import ky from 'ky';
 
 import { SOKEY_API_URL } from '@/app/(shared)/apiUrl';
-import {
-  ProductDetailRes,
-  ProductReq,
-  ProductSimilarListRes,
-  ProductsRes,
-  ProductStatusEnum,
-} from '@/app/types/api/product';
+import { ProductReq, ProductsRes, ProductStatusEnum } from '@/app/types/api/product';
 
 const getProductsQueryKey = (params?: ProductReq) => [
   'products',
@@ -37,22 +31,5 @@ export const getProductsQueryObject = (params?: ProductReq) => {
           }),
         })
         .json(),
-  };
-};
-
-// 제품 상세 조회 쿼리 객체
-export const getProductsDetailQueryObject = (productId: string) => {
-  return {
-    queryKey: ['product-detail', productId],
-    queryFn: async (): Promise<ProductDetailRes> => await ky.get(`${SOKEY_API_URL.PRODUCTS}/${productId}`).json(),
-  };
-};
-
-// 유사 제품 리스트 조회 쿼리 객체
-export const getProductSimilarListQueryObject = (productId: string) => {
-  return {
-    queryKey: ['product-similar'],
-    queryFn: async (): Promise<ProductSimilarListRes> =>
-      await ky.get(`${SOKEY_API_URL.PRODUCTS}/${productId}/similar`).json(),
   };
 };
